fix: [FX-2739] Circularity in Field and charts
