+++ packages/picasso/src/PageSidebar/index.ts
@@ -5,8 +5,4 @@ import { Props } from './PageSidebar'
 export { default } from './PageSidebar'
 export type PageSidebarProps = OmitInternalProps<Props>
 
-export {
-  SidebarContextProvider,
-  useSidebarContext
-} from './SidebarContextProvider'
 export type { Props as SidebarContextProviderProps } from './SidebarContextProvider'
