+++ packages/picasso/src/RichTextEditor/types.ts
@@ -1,3 +1,9 @@
 import { ChangeHandler } from '../QuillEditor'
 
 export type RichTextEditorChangeHandler = ChangeHandler
+
+export type CounterMessageSetter = (
+  limit: number,
+  currLength: number,
+  isError: boolean
+) => string
