+++ packages/topkit-analytics-charts/src/AnalyticsChart/AnalyticsChart.tsx
@@ -9,23 +9,7 @@ import {
   getYAxisTicks,
   formatYAxisTick
 } from './../utils'
-
-export type ChartGranularity = 'month' | 'week' | 'day' | 'hour'
-
-export type Point = {
-  id: string
-  values: Record<string, number | null>
-}
-
-export type Highlight = {
-  data: string[]
-  color: string
-}
-
-export type ReferenceLine = {
-  data: Record<string, number>
-  color: string
-}
+import { Point, Highlight, ReferenceLine, ChartGranularity } from './types'
 
 export type Props = BaseLineChartProps & {
   /**
