+++ packages/topkit-analytics-charts/src/AnalyticsChart/index.ts
@@ -1,7 +1,1 @@
 export { default } from './AnalyticsChart'
-export type {
-  Highlight,
-  Point,
-  ReferenceLine,
-  ChartGranularity
-} from './AnalyticsChart'
