+++ packages/topkit-analytics-charts/src/utils/get-x-axis-ticks/get-x-axis-ticks.ts
@@ -1,6 +1,6 @@
 import { OrderedChartDataPoint } from '@toptal/picasso-charts'
 
-import { ChartGranularity } from '../../AnalyticsChart'
+import { ChartGranularity } from '../../AnalyticsChart/types'
 
 const getXAxisTicks = (
   orderedData: OrderedChartDataPoint[],
