+++ packages/topkit-analytics-charts/src/utils/to-chart-format/to-chart-format.ts
@@ -1,6 +1,6 @@
 import { ChartDataPoint } from '@toptal/picasso-charts'
 
-import { Point, ReferenceLine } from './../../AnalyticsChart'
+import { Point, ReferenceLine } from './../../AnalyticsChart/types'
 import generateReferenceKey from '../generate-reference-key'
 
 const toChartFormat = (
